implemented status subcommand to oltp

+++ include/tateyama/util/proc_mutex.h
@@ -21,6 +21,7 @@
 
 #include <boost/filesystem/path.hpp>
 #include <boost/filesystem/operations.hpp>
+#include <boost/filesystem/fstream.hpp>
 
 namespace tateyama::server {
 
@@ -28,10 +29,20 @@ static const boost::filesystem::path LOCK_FILE_NAME = boost::filesystem::path("t
 
 class proc_mutex {
   public:
-    proc_mutex(boost::filesystem::path directory) : file_name_(directory / LOCK_FILE_NAME) {}
+    enum class lock_state : std::int32_t {
+        no_file = 0,
+        not_locked,
+        locked,
+        error,
+    };
+    
+    proc_mutex(boost::filesystem::path directory, bool create_file = true)
+        : file_name_(directory / LOCK_FILE_NAME), create_file_(create_file) {}
     ~proc_mutex() {
         close(fd_);
-        unlink(file_name_.generic_string().c_str());
+        if (create_file_) {
+            unlink(file_name_.generic_string().c_str());
+        }
     }
 
     proc_mutex(proc_mutex const& other) = delete;
@@ -39,10 +50,12 @@ class proc_mutex {
     proc_mutex(proc_mutex&& other) noexcept = delete;
     proc_mutex& operator=(proc_mutex&& other) noexcept = delete;
 
-    bool lock() {
-        if ((fd_ = open(file_name_.generic_string().c_str(), O_WRONLY | O_CREAT, S_IRUSR | S_IWUSR)) < 0) {
-            perror("open");
-            exit(-1);
+    [[nodiscard]] bool lock() {
+        if (create_file_) {
+            if ((fd_ = open(file_name_.generic_string().c_str(), O_WRONLY | O_CREAT, S_IRUSR | S_IWUSR)) < 0) {
+                perror("open");
+                exit(-1);
+            }
         }
         if (flock(fd_, LOCK_EX | LOCK_NB) == 0) {
             if (ftruncate(fd_, 0) < 0) {
@@ -61,10 +74,53 @@ class proc_mutex {
     void unlock() {
         flock(fd_, LOCK_UN);
     }
-
+    [[nodiscard]] inline std::string name() const {
+        return file_name_.generic_string();
+    }
+    [[nodiscard]] bool contents(std::string& str) {
+        if (check() != lock_state::locked) {
+            return false;
+        }
+        std::size_t sz = static_cast<std::size_t>(boost::filesystem::file_size(file_name_));
+        str.resize(sz, '\0');
+        boost::filesystem::ifstream file(file_name_);
+        file.read(&str[0], sz);
+        return true;
+    }
+    [[nodiscard]] lock_state check() {
+        boost::system::error_code error;
+        const bool result = boost::filesystem::exists(file_name_, error);
+        if (!result || error) {
+            return lock_state::no_file;
+        }
+        if (!boost::filesystem::is_regular_file(file_name_)) {
+            return lock_state::error;            
+        }
+        if (fd_ = open(file_name_.generic_string().c_str(), O_WRONLY); fd_ < 0) {
+            return lock_state::error;
+        }
+        if (flock(fd_, LOCK_EX | LOCK_NB) == 0) {
+            unlock();
+            return lock_state::not_locked;
+        }
+        return lock_state::locked;
+    }
+    [[nodiscard]] inline constexpr std::string_view to_string_view(lock_state value) noexcept {
+        using namespace std::string_view_literals;
+        using state = lock_state;
+        switch (value) {
+        case state::no_file: return "no_file"sv;
+        case state::not_locked: return "not_locked"sv;
+        case state::locked: return "locked"sv;
+        case state::error: return "error"sv;
+        }
+        std::abort();
+    }
+    
 private:
     boost::filesystem::path file_name_;
     int fd_{};
+    bool create_file_;
 };
 
 }  // tateyama::server;
